[feat] Use manifest manager interface - Create manifest manager corresponding to source type - Manifest manager has two method sync & clear - Clarify v1 & corev1 usage
