+++ config/release.yaml
@@ -279,7 +279,7 @@ apiVersion: apiextensions.k8s.io/v1
 kind: CustomResourceDefinition
 metadata:
   annotations:
-    controller-gen.kubebuilder.io/version: v0.4.1
+    controller-gen.kubebuilder.io/version: v0.7.0
   creationTimestamp: null
   name: applications.cd.tmax.io
 spec:
@@ -335,6 +335,14 @@ spec:
                     Source is a reference to the location of the application's
                     manifests or chart
                   properties:
+                    helm:
+                      description: Helm holds helm specific options
+                      properties:
+                        clonedRepoPath:
+                          type: string
+                        releaseName:
+                          type: string
+                      type: object
                     path:
                       description:
                         Path is a directory path within the Git repository,
@@ -392,8 +400,15 @@ spec:
                             - secretKeyRef
                           type: object
                       type: object
+                    type:
+                      description: Type specifies the type of the application's source
+                      enum:
+                        - PlainYAML
+                        - Helm
+                      type: string
                   required:
                     - repoURL
+                    - type
                   type: object
                 syncPolicy:
                   description: SyncPolicy controls when and how a sync will be performed
@@ -416,44 +431,74 @@ spec:
               description: ApplicationStatus defines the observed state of Application
               properties:
                 conditions:
-                  description: Conditions of IntegrationConfig
+                  description: Conditions of Application
                   items:
                     description:
-                      "Condition represents an observation of an object's
-                      state. Conditions are an extension mechanism intended to be used
-                      when the details of an observation are not a priori known or would
-                      not apply to all instances of a given Kind. \n Conditions should
-                      be added to explicitly convey properties that users and components
-                      care about rather than requiring those properties to be inferred
-                      from other observations. Once defined, the meaning of a Condition
-                      can not be changed arbitrarily - it becomes part of the API, and
-                      has the same backwards- and forwards-compatibility concerns of
-                      any other part of the API."
+                      "Condition contains details for one aspect of the current
+                      state of this API Resource. --- This struct is intended for direct
+                      use as an array at the field path .status.conditions.  For example,
+                      type FooStatus struct{     // Represents the observations of a
+                      foo's current state.     // Known .status.conditions.type are:
+                      \"Available\", \"Progressing\", and \"Degraded\"     // +patchMergeKey=type
+                      \    // +patchStrategy=merge     // +listType=map     // +listMapKey=type
+                      \    Conditions []metav1.Condition `json:\"conditions,omitempty\"
+                      patchStrategy:\"merge\" patchMergeKey:\"type\" protobuf:\"bytes,1,rep,name=conditions\"`
+                      \n     // other fields }"
                     properties:
                       lastTransitionTime:
+                        description:
+                          lastTransitionTime is the last time the condition
+                          transitioned from one status to another. This should be when
+                          the underlying condition changed.  If that is not known, then
+                          using the time when the API field changed is acceptable.
                         format: date-time
                         type: string
                       message:
+                        description:
+                          message is a human readable message indicating
+                          details about the transition. This may be an empty string.
+                        maxLength: 32768
                         type: string
+                      observedGeneration:
+                        description:
+                          observedGeneration represents the .metadata.generation
+                          that the condition was set based upon. For instance, if .metadata.generation
+                          is currently 12, but the .status.conditions[x].observedGeneration
+                          is 9, the condition is out of date with respect to the current
+                          state of the instance.
+                        format: int64
+                        minimum: 0
+                        type: integer
                       reason:
                         description:
-                          ConditionReason is intended to be a one-word, CamelCase
-                          representation of the category of cause of the current status.
-                          It is intended to be used in concise output, such as one-line
-                          kubectl get output, and in summarizing occurrences of causes.
+                          reason contains a programmatic identifier indicating
+                          the reason for the condition's last transition. Producers
+                          of specific condition types may define expected values and
+                          meanings for this field, and whether the values are considered
+                          a guaranteed API. The value should be a CamelCase string.
+                          This field may not be empty.
                         type: string
                       status:
+                        description: status of the condition, one of True, False, Unknown.
+                        enum:
+                          - "True"
+                          - "False"
+                          - Unknown
                         type: string
                       type:
                         description:
-                          "ConditionType is the type of the condition and
-                          is typically a CamelCased word or short phrase. \n Condition
-                          types should indicate state in the \"abnormal-true\" polarity.
-                          For example, if the condition indicates when a policy is invalid,
-                          the \"is valid\" case is probably the norm, so the condition
-                          should be called \"Invalid\"."
+                          type of condition in CamelCase or in foo.example.com/CamelCase.
+                          --- Many .condition.type values are consistent across resources
+                          like Available, but because arbitrary conditions can be useful
+                          (see .node.status.conditions), the ability to deconflict is
+                          important. The regex it matches is (dns1123SubdomainFmt/)?(qualifiedNameFmt)
+                        maxLength: 316
+                        pattern: ^([a-z0-9]([-a-z0-9]*[a-z0-9])?(\.[a-z0-9]([-a-z0-9]*[a-z0-9])?)*/)?(([A-Za-z0-9][-A-Za-z0-9_.]*)?[A-Za-z0-9])$
                         type: string
                     required:
+                      - lastTransitionTime
+                      - message
+                      - reason
                       - status
                       - type
                     type: object
