+++ pkg/apiserver/apis/v1/applications/deploy.go
@@ -22,7 +22,6 @@ import (
 	"net/http"
 
 	cdv1 "github.com/tmax-cloud/cd-operator/api/v1"
-	"github.com/tmax-cloud/cd-operator/pkg/manifestmanager"
 	"github.com/tmax-cloud/cd-operator/pkg/sync"
 	"k8s.io/apimachinery/pkg/types"
 
@@ -59,8 +58,7 @@ func (h *handler) updateDeploy(w http.ResponseWriter, req *http.Request) {
 	}
 	// sync resources with manitests
 	// TODO: application의 sync status, sync 옵션 등 추가하여 분기 필요.
-	mgr := manifestmanager.ManifestManager{Client: h.k8sClient, Context: context.Background()}
-	if err := sync.CheckSync(mgr.Client, app, true); err != nil {
+	if err := sync.CheckSync(h.k8sClient, app, true); err != nil {
 		log.Info(err.Error())
 		return
 	}
